pantalla index más bonita y base (problema de 2 id de usuario) arreglado

+++ ArriendaTuFinca/src/main/java/com/example/ArriendaTuFinca/models/Propiedad.java
@@ -45,7 +45,6 @@ public class Propiedad {
     private boolean asador;
     private int valor_noche;
     private boolean visible;
-    private int calificacion;
-
+    private float calificacion;
     Estado estado;
 }
